requirements3.11spark.txt
aiohttp==3.9.5
aiosignal==1.3.1
apache-beam==2.48.0
appdirs==1.4.4
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
atpublic==4.1.0
attrs==23.2.0
avro-python3==1.10.2
bidict==0.23.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
duckdb==1.0.0
duckdb_engine==0.13.0
exceptiongroup==1.2.1
executing==2.0.1
fastavro==1.9.4
fasteners==0.19
filelock==3.15.4
filprofiler==2023.3.1
find_libpython==0.4.0
findspark==2.0.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gcsfs==2024.6.0
google-api-core==2.19.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.64.1
hdfs==2.7.3
httplib2==0.22.0
humanize==4.9.0
ibis-framework==9.2.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
Js2Py==0.74
jsonpickle==3.2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kafka-python==2.0.2
kaleido==0.2.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
multidict==6.0.5
multipledispatch==1.0.0
nest-asyncio==1.6.0
numpy==1.24.4
oauthlib==3.2.2
objsize==0.6.1
orjson==3.10.5
packaging==23.2
pandas==2.2.2
parso==0.8.4
parsy==2.1
pemja==0.4.1
pexpect==4.9.0
pillow==10.3.0
pins==0.8.6
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
polars==0.20.31
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.23.4
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydot==1.4.2
Pygments==2.18.0
pyjsparser==2.7.1
PyJWT==2.8.0
pymongo==4.7.3
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyspark==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.6
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
snowflake-connector-python==3.11.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.0
sqlalchemy-risingwave==1.1.0
sqlalchemy-views==0.3.2
sqlglot==25.6.1
stack-data==0.6.3
tenacity==8.4.1
threadpoolctl==3.5.0
tomli==2.0.1
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zstandard==0.22.0
